extract function for rewriting ethers tx data
